Fixed category import
`annotationTool.importCategories` was undefined at the call site.

`MainView` thought it was responsible for the `importCategories` function,
setting this property on `annotationTool` in its initialization to its own version of it.
However, `importCategories` was already defined on `annotationTool`
and `MainView` in fact does not have a property of that name.

Closes #1.