Implemented basic key chord to set duration of "active annotation"
Creating a new annotation, be it a text annotation or a structured annotation,
makes that annotation the "active" annotation. Pressing the corresponding key chord
sets the duration of that annotation to the difference between its start
and the current position of the playhead, as long as that is in the future
relative to the start.