Restored and refactored `localStorage`-only models
Instead of having a list of models that should only be persisted locally in the configuration,
which was always empty in all of our examples, the models can now decide that for themselves.
This was effectively already the case before, since the `Loop` model and collection just enlisted entered themselves
into the aforementioned list without consulting any configuration keys.