Updated underscore
I had it up to here with not being able to use "newer" (< 5 years old that is) underscore features.
Besides, underscore should be a relatively safe library to update. It had a few bigger compatibility issues in the past,
but is generally pretty stable and a casual perusal of the changelog brought up nothing that should affect us.

I also tested the app a bit after this, but only manually, so take this with a grain of salt.
Whenever things break in the future, you might consider underscore as a possible cause.

Oh if only we had automated tests ...