Updated `lessc` to actually support source maps
This necessitated a few changes to our ancient version of Bootstrap,
which is suboptimal of course, but I figured/hoped that it, too,
will be updated at some point and also managed with
some kind of package manager.
Until then, we might as well make use of the fact
that the code of this dependency is our under full control.

Coincidentally, this also fixes a long standing CSS bug in the timeline. :)