Switched to non-minified Backbone for now
Since Backbone is our UI framework, most of the stacktraces I see
contain many layers of Backbone. These traces contain
such readable function names as `d` or `g`.
You can't debug like this.

Of course it would be nicer to have source maps,
but hunting for the right version and then just stapling it on
seems wrong, too.
If we have a source map, it should come with the dependency,
optimally through some kind of dependency manager ...

Also this should actually not matter for the production build,
since we send everything through `r.js` in the end anyway.