Removed a useless Handlebars helper
While there is actually an XSS risk in the places where this helper was used,
it did nothing to prevent the possible attacks.
Handlebars escapes HTML special characters by default,
however this is not the danger here.

This removes the false sense of security this offers
and also gets rid of the complexity that it brings.