Removed redundant access specifiers from track related code
Specifically, code that creates tracks, passed around an access level
in addition to the one carried by the base resource.
This is redundant and confusing
and was actually the source of a bug in the tests
that was covered up by them always being run
with a user that has admin priviliges,
and can thus see everything.
The latter part is about to change, though,
which made me notice this.