Updated Bootstrap to a slightly less outdated version ...
Also used this opportunity to fix some things in the dependencies
and RequireJS configuration to do with that.

I tried to recreate all the layout and style things I saw breaking
but I would not bet my life on this being 100% faithful.
But since the CSS is pretty janky anyway, we should just fix
the things we care about as we go along.