Update jQuery
I would love to use Promises to coordinate the asynchronous parts
of the tool more, but the ancient jQuery we use makes this
nigh impossible. This small upgrade at least moves the semantics
of jQueries own broken implementation of Promises/Deferreds
a bit closer to the spec.

I also removed the minified version, which was unused anyway.
Minification, if it is needed, should (in the current setup at least)
be handled by the RequireJS grunt task. In the long run,
we of course want proper dependency management and source maps. :)