Correctly initialize template layout settings
When loading a layout from a template,
it does not have a `settings` key,
so extending it has no effect.

This results in the maximize and pop-out icons being shown
when you load a predefined view,
although these should be hidden,
since we currently don't support these GoldenLayout features,
unfortunately. :(