Get rid of JSDoc `@requires` declarations
These are always out of sync with what the module actually requires,
and I'm not spending any time maintaining these.
I also don't see any need for them.
If you want them back, you are welcome to implement a way
to add them automatically, or wait for me to get around to it. ;)