Fix showing previously invisible tracks
The problem was that we did not listen for the `"reset"` event
on the annotation collection in the timeline, which is triggered
when the annotations are fetched for the server.

This also contains some minor refactorings around the relevant code,
that makes this simpler. It also stops `Track` from saving attributes
like `"visible"` and `"annotationsLoaded"` in the `localStorage`
in development mode, which could lead to some strange behavior before.