Use Maven to clean the frontend artifacts in the `clean` phase
The `node` and `npm` installation tasks are not run
during Mavens normal `clean` lifecycle.
Neither is `npm install`.
So if you run `mvn clean install` in a clean copy of the repo
like we do in the CI scripts
the call to `grunt clean` will fail.

We could of course just add these installation tasks
to the `clean` lifecycle as well, but that seems kind of wasteful.