Tell `Categories` collection when changing `id`-s in `Category#sync`
`Backbone.Collection`-s maintain an index of their `Model`-s
by their `id`. If we change that, we need to go through the proper
Backbone channels, so that the `Collection` has an opportunity
to update it. Otherwise, subsequent operations might fail.

The surprising thing to note here is that the `model` argument
to `sync` is not actually a copy of the `Model`-s `attributes`
or anything like that. It is the thing itself.
So after a call to `Category#sync`, we permanently changed
its `id`. This is fine, apparently, but not the cleanest solution.
Maybe we want to refactor this a bit more down the line.

Closes #500.