Fix the backend tests
The `contentType` calls are apparently necessary because
when you don't specify a `formParam` in a call to REST Assured,
it defaults to a `ContentType` of something other than
`application/x-www-form-urlencoded`, which the JAX-RS stuff
refuses to answer in the presence of `@FormParameter` annotations.

This was different when we still mostly dependet/ran against
Opencast 8, so I'm thinking something in the REST test environment
we use changed. ¯\_(ツ)_/¯

(cherry picked from commit c1416aecbcc88ad45e0640332b2626025bc18818)