Fixing potential raise condition when saving categories
Sometimes when editing a category which is attached to a series and
then making this category local and saving it, the category labels get
duplicated. This might have to do with the asynchronous JavaScript
request for the category and its labels.

This patch reduces the number of requests sent, plus makes the save
operation for the category itself synchronous, ensuring the label
modifications happen after the operation is actually done.