Fix db compatibility issues with Opencast 13
Opencast 13 reworked some of their common db classes (in PR opencast#3903),
which we also use in the Annotation Tool, thus breaking it.

This commit attempts to fix the breaking changes by using the new classes
introduced in Opencast 13. From a user perspective, this should change
exactly nothing. There is also no DB migration, as the db entries
themselves have not changed.

This commit is based on groundwork by @KatrinIhler, massive thanks to her.