Get rid of warnings regarding parallel builds
When building with Maven's `-T` option the build was warning
about two plugins that might not be thread safe.

This patch updates one and removes the other.

Regarding static weaving: It might be interesting to look at
at some point, but it's "only" for performance and Opencast removed
this as well: https://github.com/opencast/opencast/pull/3031.