Hide the fact that we are duplicating series labels better
Seriously, though: The way series categories work with MCA is broken; cf. #648.
This fixes a superficial symptom of that fact, namely #634,
but it doesn't address the underlying problem of infinitely duplicating labels.

The problem with the old code was that the deleted labels are still included
in the response in their undeleted form.
This patch just updates them in the response collection in-place.